backend/pyproject.toml
[build-system]
requires = ["setuptools>=65.0"]
build-backend = "setuptools.build_meta"

[project]
name = "vibeanalytix"
version = "0.1.0"
description = "AI-powered code understanding engine with deliberate reasoning pipeline"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "sqlalchemy[asyncio]==2.0.23",
    "asyncpg==0.29.0",
    "alembic==1.13.1",
    "celery[redis]==5.3.4",
    "redis==5.0.1",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    "python-multipart==0.0.6",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "python-dotenv==1.0.0",
    "tree-sitter==0.20.4",
    "tree-sitter-python==0.20.4",
    "tree-sitter-javascript==0.21.0",
    "tree-sitter-typescript==0.21.0",
    "tree-sitter-java==0.21.0",
    "tree-sitter-go==0.21.0",
    "tree-sitter-c==0.21.0",
    "tree-sitter-cpp==0.21.0",
    "openai==1.3.6",
    "pgvector==0.2.4",
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "hypothesis==6.92.0",
]

[project.optional-dependencies]
dev = [
    "pytest-cov==4.1.0",
    "black==23.12.0",
    "ruff==0.1.8",
    "mypy==1.7.0",
]
